Fixed collisions
X: The check for whether we were "crossing an even number" was wrong. The check we made checked for crosses uneven<->even, when it should have been even<->uneven.
Y: Confused the scenarios. Hitting the top of the brick VS hitting the bottom was swapped due to forgetting that the positive Y direction is down.